:memo: ContextValue (refactoring and docs)
