+++ src/test/java/io/github/syst3ms/skriptparser/syntax/EvtTest.java
@@ -5,7 +5,7 @@ import io.github.syst3ms.skriptparser.lang.Expression;
 import io.github.syst3ms.skriptparser.lang.SkriptEvent;
 import io.github.syst3ms.skriptparser.lang.TriggerContext;
 import io.github.syst3ms.skriptparser.parsing.ParseContext;
-import io.github.syst3ms.skriptparser.registration.contextvalues.ContextValueInfo.Usage;
+import io.github.syst3ms.skriptparser.registration.contextvalues.ContextValue.Usage;
 import io.github.syst3ms.skriptparser.syntax.TestContext.SubTestContext;
 
 import java.time.Duration;
